update (SMAPI 4.0.4, SDV 1.6.3)

+++ source/~MindMeltMax/Quality Bait/Config.cs
@@ -9,11 +9,6 @@
 *************************************************/
 
 using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace QualityBait
 {
@@ -22,6 +17,6 @@ namespace QualityBait
         public int ChancePercentage { get; set; } = 75;
 
         [JsonIgnore]
-        public double Chance => (double)ChancePercentage / 100.0;
+        public double Chance => ChancePercentage / 100.0;
     }
 }
